+++ lib/main.dart
@@ -4,7 +4,6 @@ import 'package:smart_home/blocs/blocs.dart';
 import 'package:smart_home/blocs/simple_bloc_delegate.dart';
 import 'package:smart_home/pages/home_page.dart';
 import 'package:smart_home/pages/login_page.dart';
-import 'package:smart_home/pages/search_page.dart';
 
 void main() {
   BlocSupervisor.delegate = SimpleBlocDelegate();
@@ -28,12 +27,14 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      routes: {
-        '/': (context) => HomePage(),
-        '/login': (context) => LoginPage(),
-        '/search': (context) => SearchPage(),
-      },
-      initialRoute: '/',
+      title: '智慧家庭',
+      home: BlocBuilder<AuthenticationBloc, AuthenticationState>(
+          builder: (context, state) {
+            if (state is Authenticated) {
+              return HomePage() ;
+            }
+        return LoginPage();
+      }),
     );
   }
 }
