+++ lib/pages/home_page.dart
@@ -1,40 +1,37 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:smart_home/blocs/blocs.dart';
+import 'package:smart_home/pages/search_page.dart';
 
 class HomePage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return BlocListener<AuthenticationBloc, AuthenticationState>(
-      listener: (context, state) {
-        if (state is Unauthenticated) {
-          Navigator.of(context).pushNamed('/login');
-        }
-      },
-      child: Scaffold(
-        appBar: AppBar(
-          title: Text('首页'),
-        ),
-        body: Center(
-          child: Column(
-            children: <Widget>[
-              FlatButton(
-                child: Text('注销'),
-                onPressed: () {
-                  BlocProvider.of<AuthenticationBloc>(context)
-                      .add(AuthenticationLogout());
-                },
-              ),
-              FlatButton(
-                child: Text('搜索'),
-                onPressed: () {
-                  Navigator.of(context).pushNamed('/search');
-                  BlocProvider.of<StorageBloc>(context)
-                      .add(StorageSearchStarted());
-                },
-              )
-            ],
-          ),
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('首页'),
+      ),
+      body: Center(
+        child: Column(
+          children: <Widget>[
+            FlatButton(
+              child: Text('注销'),
+              onPressed: () {
+                BlocProvider.of<AuthenticationBloc>(context)
+                    .add(AuthenticationLogout());
+              },
+            ),
+            FlatButton(
+              child: Text('搜索'),
+              onPressed: () {
+                BlocProvider.of<StorageBloc>(context)
+                    .add(StorageSearchStarted());
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(builder: (context) => SearchPage()),
+                );
+              },
+            )
+          ],
         ),
       ),
     );
