+++ lib/pages/login_page.dart
@@ -1,21 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:smart_home/blocs/authentication/authentication_bloc.dart';
 import 'package:smart_home/widgets/login_form.dart';
 
 class LoginPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return BlocListener<AuthenticationBloc, AuthenticationState>(
-      listener: (context, state) {
-        if (state is Authenticated) {
-          Navigator.of(context).pop();
-        }
-      },
-      child: Scaffold(
-        appBar: AppBar(title: Text("登录")),
-        body: LoginForm(),
-      ),
+    return Scaffold(
+      appBar: AppBar(title: Text("登录")),
+      body: LoginForm(),
     );
   }
 }
